requirements.txt
Django
influxdb
ipdb
ipython
requests
tweepy
websocket-client
colorama
emoji